pyproject.toml
[tool.poetry]
name = "discord-sneakers-bot"
version = "0.1.0"
description = ""
authors = ["pedro.grilo"]
readme = "README.md"
packages = [{include = "discord_sneakers_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.30.0"
beautifulsoup4 = "^4.12.2"
"discord.py" = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
